Formatクラスの削除
