+++ 08.ls_object/lib/command.rb
@@ -1,23 +1,33 @@
 # frozen_string_literal: true
 
 require_relative 'file_stat'
-require_relative 'format'
+require_relative 'long_renderer'
+require_relative 'short_renderer'
 
 class Command
-  def initialize(path, params)
-    @file_stats = generate_file_stats_classes(path)
-    @params = params
+  def initialize(path, all: false, reverse: false, long: false)
+    @file_stats = generate_file_stats_classes(path, all, reverse)
+    @long = long
   end
 
   def run
-    format = Format.new(@file_stats, **@params)
-    format.render
+    renderer = @long ? LongRenderer.new(@file_stats) : ShortRenderer.new(@file_stats)
+    renderer.render
   end
 
   private
 
-  def generate_file_stats_classes(path)
+  def generate_file_stats_classes(path, all, reverse)
     paths = Dir.foreach(path).map { |entry| File.join(path, entry) }
-    paths.map { |p| FileStat.new(p) }
+    file_stats = paths.map { |p| FileStat.new(p) }
+    file_stats.sort_by! { |f| ignore_dot_in_dotfile(f.name, f.dotfile) }
+    file_stats.reverse! if reverse
+    file_stats.reject! { |f| /^[.]+/.match?(f.name) } unless all
+
+    file_stats
+  end
+
+  def ignore_dot_in_dotfile(name, dotfile)
+    dotfile ? name.delete_prefix('.') : name
   end
 end
